Port to timsrust 0.6.6 via a local 0.4.2-compatible adapter
Depend on timsrust-tdf 0.6.6 (TDF reader only) and rusqlite 0.35. New
src/tsr.rs keeps 0.4.2's positional frame order, fractional Tof2Mz and
Scan2Im converters, and GlobalMetadata parsing; call sites, examples and
tests switch to crate::tsr. tests/targeted.rs casts unsigned SQL params.

+++ tests/targeted.rs
@@ -32,7 +32,7 @@ fn replace_points(input: &Path, changes: &[(usize, Points)]) {
         let record = dnoise::codec::encode_frame_type2(*scans, points);
         file.write_all(&record).unwrap();
         db.execute("UPDATE Frames SET TimsId=?1,NumPeaks=?2,MaxIntensity=?3,SummedIntensities=?4 WHERE Id=?5",
-            rusqlite::params![offset,points.len(),points.iter().map(|p|p.2).max().unwrap_or(0),points.iter().map(|p|p.2 as u64).sum::<u64>(),index+1]).unwrap();
+            rusqlite::params![offset as i64,points.len() as i64,points.iter().map(|p|p.2).max().unwrap_or(0),points.iter().map(|p|p.2 as u64).sum::<u64>() as i64,(index+1) as i64]).unwrap();
         offset += record.len();
     }
 }
